fix: flyway migration 충돌 수정

+++ backend/src/main/resources/db/migration/prod/V31__create_keyword_table.sql
@@ -3,8 +3,8 @@ CREATE TABLE keyword
     id          BIGINT       NOT NULL AUTO_INCREMENT,
     name        VARCHAR(255) NOT NULL,
     description VARCHAR(255) NOT NULL,
-    seq         BIGINT       NOT NULL,
-    importance  BIGINT       NOT NULL,
+    seq         INT      NOT NULL,
+    importance  INT      NOT NULL,
     session_id  BIGINT       NOT NULL,
     parent_id   BIGINT,
     CONSTRAINT pk_keyword PRIMARY KEY (id)
